A massive refactor, I did outside of GIT version control.
ZOMG! jQuery pong!! W00t, yeah.
Based on the one by a guy named
@author Ben White <benwhite@columbus.rr.com>
@author Ben Ogle <ben@benogle.com>

I have massively refactored the code for this.  I added sound effects,
fonts, score and a background image.  I'm trying to make it look as
much like the original pong as possible.  I managed to remove the need
for a ball image by taking advantage of CSS3 rounded corners, but
they're an option.

I will be working on a complete re-write of this soon, as it needs to
be built again from the ground up.